Move `Range` to the `time` module
`Range` is a time span, not a pipeline concept, and `track` needs it
without depending on `pipeline`. It stays re-exported from `pipeline`.
