+++ crates/motiongfx/src/time.rs
@@ -38,6 +38,29 @@ pub const fn ns(nanos: u64) -> Duration {
     Duration::from_nanos(nanos)
 }
 
+/// A half-open time span `[start, end)`.
+#[derive(Default, Debug, PartialEq, Eq, Clone, Copy)]
+pub struct Range {
+    pub start: Duration,
+    pub end: Duration,
+}
+
+impl Range {
+    /// True when the two [`Range`]s share any point, including a
+    /// boundary-only touch.
+    pub fn overlap(&self, other: &Self) -> bool {
+        self.start <= other.end && other.start <= self.end
+    }
+
+    /// The span shared by both [`Range`]s, or `None` when they share
+    /// no span of positive length. A boundary-only touch is `None`.
+    pub fn intersect(&self, other: &Self) -> Option<Self> {
+        let start = self.start.max(other.start);
+        let end = self.end.min(other.end);
+        (start < end).then_some(Self { start, end })
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -48,5 +71,32 @@ mod tests {
         assert_eq!(cs(150), ms(1_500));
         assert_eq!(ms(1), ns(1_000_000));
         assert_eq!(cs(u64::MAX), Duration::from_millis(u64::MAX));
+    }
+
+    #[test]
+    fn range_overlap_counts_boundary_touch() {
+        let a = Range { start: s(0), end: s(5) };
+        let b = Range { start: s(3), end: s(8) };
+        let c = Range { start: s(6), end: s(10) };
+        let touching = Range { start: s(5), end: s(5) };
+
+        assert!(a.overlap(&b));
+        assert!(!a.overlap(&c));
+        assert!(a.overlap(&touching));
+    }
+
+    #[test]
+    fn range_intersect_drops_boundary_touch() {
+        let a = Range { start: s(0), end: s(5) };
+        let b = Range { start: s(3), end: s(8) };
+        let c = Range { start: s(6), end: s(10) };
+        let touching = Range { start: s(5), end: s(8) };
+
+        assert_eq!(
+            a.intersect(&b),
+            Some(Range { start: s(3), end: s(5) }),
+        );
+        assert_eq!(a.intersect(&c), None);
+        assert_eq!(a.intersect(&touching), None);
     }
 }
